refactor: mandara.jsを分割して1000行以下に
- mandara-list-view.jsを作成してリストビュー機能を分離
- 重複コードを削除してヘルパー関数に統合
- mandara.js: 978行（元1166行から188行削減）
- mandara-list-view.js: 207行
